Use pool vaa nonce for eth and aptos

+++ ethereum/contracts/omnipool/WormholeAdapterPool.sol
@@ -28,7 +28,7 @@ contract WormholeAdapterPool {
 
     // todo! Delete after wormhole running
     mapping(uint32 => bytes) public cachedVAA;
-    uint32 vaaNonce;
+    uint64 vaaNonce;
 
     event PoolWithdrawEvent(
         uint64 nonce,
@@ -78,7 +78,7 @@ contract WormholeAdapterPool {
         //                encodedVm
         //            );
         LibPoolCodec.ManagePoolPayload memory payload = LibPoolCodec
-            .decodeManagePoolPayload(encodedVm);
+        .decodeManagePoolPayload(encodedVm);
         require(
             payload.poolCallType == LibPoolCodec.POOL_REGISTER_OWNER,
             "INVALID CALL TYPE"
@@ -96,7 +96,7 @@ contract WormholeAdapterPool {
         //                encodedVm
         //            );
         LibPoolCodec.ManagePoolPayload memory payload = LibPoolCodec
-            .decodeManagePoolPayload(encodedVm);
+        .decodeManagePoolPayload(encodedVm);
         require(
             payload.poolCallType == LibPoolCodec.POOL_DELETE_OWNER,
             "INVALID CALL TYPE"
@@ -114,7 +114,7 @@ contract WormholeAdapterPool {
         //                encodedVm
         //            );
         LibPoolCodec.ManagePoolPayload memory payload = LibPoolCodec
-            .decodeManagePoolPayload(encodedVm);
+        .decodeManagePoolPayload(encodedVm);
         require(
             payload.poolCallType == LibPoolCodec.POOL_REGISTER_SPENDER,
             "INVALID CALL TYPE"
@@ -132,7 +132,7 @@ contract WormholeAdapterPool {
         //                encodedVm
         //            );
         LibPoolCodec.ManagePoolPayload memory payload = LibPoolCodec
-            .decodeManagePoolPayload(encodedVm);
+        .decodeManagePoolPayload(encodedVm);
         require(
             payload.poolCallType == LibPoolCodec.POOL_DELETE_SPENDER,
             "INVALID CALL TYPE"
@@ -158,13 +158,13 @@ contract WormholeAdapterPool {
             LibAsset.maxApproveERC20(IERC20(token), address(dolaPool), amount);
         }
 
-        bytes memory payload = dolaPool.deposit{value: msg.value - wormholeFee}(
+        bytes memory payload = dolaPool.deposit{value : msg.value - wormholeFee}(
             token,
             amount,
             appId,
             appPayload
         );
-        wormhole.publishMessage{value: wormholeFee}(
+        wormhole.publishMessage{value : wormholeFee}(
             0,
             payload,
             wormholeFinality
@@ -176,13 +176,13 @@ contract WormholeAdapterPool {
 
     /// Send message that do not involve incoming or outgoing funds by application
     function sendMessage(uint16 appId, bytes memory appPayload)
-        external
-        payable
+    external
+    payable
     {
         uint256 wormholeFee = wormhole.messageFee();
         require(msg.value >= wormholeFee, "FEE NOT ENOUGH");
         bytes memory payload = dolaPool.sendMessage(appId, appPayload);
-        wormhole.publishMessage{value: msg.value}(0, payload, wormholeFinality);
+        wormhole.publishMessage{value : msg.value}(0, payload, wormholeFinality);
         cachedVAA[getNonce()] = payload;
         increaseNonce();
     }
@@ -197,7 +197,7 @@ contract WormholeAdapterPool {
         //                encodedVm
         //            );
         LibPoolCodec.WithdrawPayload memory payload = LibPoolCodec
-            .decodeWithdrawPayload(encodedVm);
+        .decodeWithdrawPayload(encodedVm);
         dolaPool.withdraw(payload.user, payload.amount, payload.pool);
 
         emit PoolWithdrawEvent(
